[PATCH authentication] add missing analytics properties for social auth
