+++ src/Components/Authentication/Desktop/ModalManager.tsx
@@ -1,5 +1,4 @@
 import { FormikProps } from "formik"
-import qs from "querystring"
 import React, { Component } from "react"
 
 import { DesktopModal } from "Components/Authentication/Desktop/Components/DesktopModal"
@@ -28,6 +27,7 @@ export interface ModalManagerProps {
     formikBag: FormikProps<InputValues>
   ) => void
   blurContainerSelector?: string
+  onSocialAuthEvent?: (options) => void
 }
 
 export interface ModalManagerState {
@@ -111,28 +111,6 @@ export class ModalManager extends Component<
       ? this.props.handleSubmit.bind(this, currentType, options)
       : defaultHandleSubmit(submitUrls[currentType], csrf, redirectTo)
 
-    const queryData = Object.assign(
-      {},
-      options,
-      {
-        accepted_terms_of_service: true,
-        agreed_to_receive_emails: true,
-        "signup-referer": options.signupReferer || location.href,
-      },
-      options.redirectTo
-        ? {
-            "redirect-to": options.redirectTo,
-          }
-        : null,
-      options.intent
-        ? {
-            "signup-intent": options.intent,
-          }
-        : null
-    )
-
-    const authQueryData = qs.stringify(queryData)
-
     return (
       <DesktopModal
         blurContainerSelector={blurContainerSelector}
@@ -146,14 +124,10 @@ export class ModalManager extends Component<
           type={currentType}
           error={error}
           handleSubmit={handleSubmit}
-          onFacebookLogin={() =>
-            (window.location.href = submitUrls.facebook + `?${authQueryData}`)
-          }
-          onTwitterLogin={() =>
-            (window.location.href = submitUrls.twitter + `?${authQueryData}`)
-          }
+          submitUrls={submitUrls}
           options={options}
           handleTypeChange={this.handleTypeChange}
+          onSocialAuthEvent={this.props.onSocialAuthEvent}
         />
       </DesktopModal>
     )
